fix: use uint16_t for bytecode to prevent 8-bit overflow
The instruction list grew to 255 instructions, so when a while loop tried to reference the end of the currently generated list, it received 256 which it interpreted as 0. This then caused the loop to skip to the start of the instruction list, which makes it start all over, and never complete the condition.

The instructions now use uint16_t, so we will encounter this issue again when we reach a list size of 65535. I added a note to read later regarding moving towards size_t. I didn't want to do this right now without benchmarking the performance difference.

I considered adding an unsigned type to Value and making the bytecode just reference a constant value type, but I had two key issues with that:
1. indirection cost
2. i will have the same issue when the number of constants reaches over 255. Will I use another indirection then?

For this reason, I decided to keep the Values as-is and just given the bytecode more memory.

+++ libs/runtime/include/tachyon/runtime/bytecode.h
@@ -21,7 +21,7 @@ namespace tachyon::runtime
  * | 0xF0  | intrinsic       |
  * +-------+-----------------+
  */
-enum Bytecode : uint8_t
+enum Bytecode : uint16_t
 {
     /* machine */
     RETV = 0x00, // return void
